--- src/connect-dai-wallet/src/hooks/useModal.js
@@ -1,11 +1,0 @@
-import { useContext } from 'react';
-import { ModalStateContext } from 'providers/ModalProvider';
-
-function useModal() {
-  const context = useContext(ModalStateContext);
-  const { show, reset } = context;
-  const showByType = modalType => show({ modalType });
-  return { show, reset, showByType };
-}
-
-export default useModal;
