+++ src/connect-dai-wallet/src/utils/constants.js
@@ -1,18 +1,3 @@
-export const CDP_SAFETY_LEVELS = {
-  DANGER: 'danger',
-  WARNING: 'warning',
-  NEUTRAL: 'neutral',
-  SAFE: 'safe'
-};
-
-export const TxLifecycle = {
-  NULL: 'null',
-  INITIALIZED: 'initialized',
-  PENDING: 'pending',
-  MINED: 'mined',
-  ERROR: 'error'
-};
-
 export const AccountTypes = {
   LEDGER: 'ledger',
   TREZOR: 'trezor',
